refactor(utils): format file size
