+++ website/app/components/tools/ResultsTable.tsx
@@ -1,5 +1,6 @@
 import { ActionIcon, Button, Stack, Table, Text, Title } from '@mantine/core';
 import { IconFileDownload } from '@tabler/icons-react';
+import { formatFileSize } from './utils';
 
 interface ConversionResult {
 	name: string;
@@ -14,18 +15,6 @@ interface ResultsTableProps {
 	downloadError?: string | null;
 	disabled?: boolean;
 }
-
-const formatFileSize = (bytes: number): string => {
-	if (bytes === 0) {
-		return '0 bytes';
-	}
-
-	const k = 1024;
-	const sizes = ['bytes', 'KB', 'MB', 'GB'];
-
-	const i = Math.floor(Math.log(bytes) / Math.log(k));
-	return `${Number.parseFloat((bytes / k ** i).toFixed(2))} ${sizes[i]}`;
-};
 
 export const ResultsTable = ({
 	results,
